w2layout: strictly check that content.render is a function when this attribute is available. When this method is not provided, print the content through w2utils.render() to ensure that any 'unexpected type' also is presented by a legible content string (the result is immediately noticable in layout example No.8 (layout with tabs) as previously clicking on any tab would render *nothing* as the w2layout.content() interface is fed an *event* *object* straight away, which itself would produce an empty string.
Of course one may argue that placing the w2utils.render() call inside w2layout.content() is wrong and should have been placed in the one offending example (layout 8) but I have found time and again that it is hard to analyze the cause of trouble when a w2layout panel shows up unexpectedly empty; generally those issues are due to misbehaving user code but I like w2ui to help me at least noticing such mistakes as soon and as clearly as possible. w2utils.render() is designed to help that goal and that is the reason why the w2ui library code itself, in this case w2layout, merits the inclusion of this call, rather than duplicating it all over the user code - and probably missing one or more spots where it would have saved us hassle.